Allow customization of standard checks run on /health_check
Previously the standard checks were hard coded inside of the utils
checker. The specific use case was an interest in disabling the
migrations check by default, so that instances would not be kicked out
of the load balancer if the migration version was different. Obviously
that may not be what everyone wants, so this allows easy customization.