Drop rails 4.0 support
Rails <4.2 is deprecated and no longer receives security fixes.
Dropping support for 4.0 and even 4.1 will likely not affect many users of
the latest code.

Currently, the latest version of turbolinks will install against the
current test suite, causing failures in the 4.0 set of tests, because the
changes made are incompatible with 4.0. This has been causing failures
on pull requests for at least a month.