Use ENV['REDIS_PASSWORD'] as default password for redis
When merging #103 into master in 6a4b72c3534483d311e373f64af6dcc6f9d870b5, the redis password was accidentally set to `redis_password`. This leads to failures in environments without password protection. Furthermore, the README suggests that `ENV['REDIS_PASSWORD']` would be used as a default. This change removes the incorrect default and replaces it with the documented one.